Making clinical document subclass optional
Athena allows for clinical documents without a subclass, both in the UI,
as well as from [the documentation](https://docs.athenahealth.com/api/workflows/document-classification-guide)

> Subclassification is not necessary for [clinical documents]

However, because it's Athena, the API still requires the documentsubclass
field to be set.

It looks like the original author already discovered the nasty magic
string CLINICALDOCUMENT as a "default" option for the subclass for
clinical documents. This change makes it so the end user doesn't have to
worry about the magic string.